simulate processed events
To show what the end data schema is intended to look like, I wrote a function to flatten the JSON structure of general CDEvent payloads and join them together into a tabular format.  In S3, they will ideally be kept in separate files and stored in columnar format (Parquet), but for ease of use and practice with dashboards, I figured this was the best option for now.